Move debug folder out of extension folder
