+++ src/debug/dart_debug_impl.ts
@@ -3,16 +3,16 @@ import * as fs from "fs";
 import * as path from "path";
 import { DebugSession, Event, InitializedEvent, OutputEvent, Scope, Source, StackFrame, StoppedEvent, TerminatedEvent } from "vscode-debugadapter";
 import { DebugProtocol } from "vscode-debugprotocol";
-import { VmServiceCapabilities } from "../../shared/capabilities/vm_service";
-import { observatoryListeningBannerPattern, pleaseReportBug } from "../../shared/constants";
-import { LogCategory, LogSeverity } from "../../shared/enums";
-import { LogMessage } from "../../shared/interfaces";
-import { flatMap, throttle, uniq, uriToFilePath } from "../../shared/utils";
-import { sortBy } from "../../shared/utils/array";
-import * as col from "../../shared/utils/colors";
-import { config } from "../config";
-import { getLogHeader } from "../utils/log";
-import { safeSpawn } from "../utils/processes";
+import { config } from "../extension/config";
+import { getLogHeader } from "../extension/utils/log";
+import { safeSpawn } from "../extension/utils/processes";
+import { VmServiceCapabilities } from "../shared/capabilities/vm_service";
+import { observatoryListeningBannerPattern, pleaseReportBug } from "../shared/constants";
+import { LogCategory, LogSeverity } from "../shared/enums";
+import { LogMessage } from "../shared/interfaces";
+import { flatMap, throttle, uniq, uriToFilePath } from "../shared/utils";
+import { sortBy } from "../shared/utils/array";
+import * as col from "../shared/utils/colors";
 import { DebuggerResult, ObservatoryConnection, SourceReportKind, Version, VM, VMClass, VMClassRef, VMErrorRef, VMEvent, VMFrame, VMInstance, VMInstanceRef, VMIsolate, VMIsolateRef, VMLibrary, VMMapEntry, VMObj, VMScript, VMScriptRef, VMSentinel, VMSourceReport, VMStack, VMTypeRef } from "./dart_debug_protocol";
 import { DebugAdapterLogger } from "./logging";
 import { PackageMap } from "./package_map";
