+++ src/debug/dart_debug_protocol.ts
@@ -1,5 +1,5 @@
 import * as WebSocket from "ws";
-import { PromiseCompleter } from "../../shared/utils";
+import { PromiseCompleter } from "../shared/utils";
 
 export class DebuggerResult {
 	public result: VMResponse;
