+++ src/debug/flutter_web_run.ts
@@ -1,4 +1,4 @@
-import { Logger } from "../../shared/interfaces";
+import { Logger } from "../shared/interfaces";
 import { FlutterRunBase, RunMode } from "./flutter_run_base";
 
 export class FlutterWebRun extends FlutterRunBase {
