+++ src/debug/logging.ts
@@ -1,8 +1,8 @@
 import { Event } from "vscode-debugadapter";
 import { DebugProtocol } from "vscode-debugprotocol";
-import { LogCategory, LogSeverity } from "../../shared/enums";
-import { Logger, LogMessage, SomeError } from "../../shared/interfaces";
-import { errorString } from "../../shared/utils";
+import { LogCategory, LogSeverity } from "../shared/enums";
+import { Logger, LogMessage, SomeError } from "../shared/interfaces";
+import { errorString } from "../shared/utils";
 
 // A logger that passes log events back to the UI in `dart.log` events.
 export class DebugAdapterLogger implements Logger {
