+++ src/debug/package_map.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import * as path from "path";
-import { isWin } from "../../shared/constants";
-import { findFile, isWithinPath, uriToFilePath } from "../../shared/utils";
+import { isWin } from "../shared/constants";
+import { findFile, isWithinPath, uriToFilePath } from "../shared/utils";
 
 export class PackageMap {
 	public static findPackagesFile<T extends string | undefined>(entryPoint: T): string | (undefined extends T ? undefined : never) {
