+++ src/debug/utils.ts
@@ -1,5 +1,5 @@
 import { DebugProtocol } from "vscode-debugprotocol";
-import { forceWindowsDriveLetterToUppercase } from "../../shared/utils";
+import { forceWindowsDriveLetterToUppercase } from "../shared/utils";
 
 export function formatPathForVm(file: string): string {
 	// Handle drive letter inconsistencies.
