+++ src/extension/providers/debug_config_provider.ts
@@ -4,6 +4,13 @@ import * as path from "path";
 import * as vs from "vscode";
 import { CancellationToken, DebugConfiguration, DebugConfigurationProvider, ProviderResult, Uri, window, workspace, WorkspaceFolder } from "vscode";
 import { DebugSession } from "vscode-debugadapter";
+import { DartDebugSession } from "../../debug/dart_debug_impl";
+import { DartTestDebugSession } from "../../debug/dart_test_debug_impl";
+import { FlutterDebugSession } from "../../debug/flutter_debug_impl";
+import { FlutterTestDebugSession } from "../../debug/flutter_test_debug_impl";
+import { FlutterWebDebugSession } from "../../debug/flutter_web_debug_impl";
+import { FlutterWebTestDebugSession } from "../../debug/flutter_web_test_debug_impl";
+import { FlutterLaunchRequestArguments } from "../../debug/utils";
 import { FlutterCapabilities } from "../../shared/capabilities/flutter";
 import { CHROME_OS_VM_SERVICE_PORT, dartVMPath, debugAnywayAction, flutterPath, HAS_LAST_DEBUG_CONFIG, isChromeOS, pubPath, pubSnapshotPath, showErrorsAction } from "../../shared/constants";
 import { Device } from "../../shared/flutter/daemon_interfaces";
@@ -15,13 +22,6 @@ import { Analytics } from "../analytics";
 import { LastDebugSession } from "../commands/debug";
 import { isLogging } from "../commands/logging";
 import { config } from "../config";
-import { DartDebugSession } from "../debug/dart_debug_impl";
-import { DartTestDebugSession } from "../debug/dart_test_debug_impl";
-import { FlutterDebugSession } from "../debug/flutter_debug_impl";
-import { FlutterTestDebugSession } from "../debug/flutter_test_debug_impl";
-import { FlutterWebDebugSession } from "../debug/flutter_web_debug_impl";
-import { FlutterWebTestDebugSession } from "../debug/flutter_web_test_debug_impl";
-import { FlutterLaunchRequestArguments } from "../debug/utils";
 import { locateBestProjectRoot } from "../project";
 import { PubGlobal } from "../pub/global";
 import { WebDev } from "../pub/webdev";
